Split the main jumpstart function
The game crashes on campaign start if both jumpstart and the race/gender aware
namelist mod are both installed and the jumpstart config is complex enough.
This was due to accidentally tripping the internal loop counter in Unreal due
to the sheer number of instructions executed.

The fix is to split up JumpStart into multiple functions, specifically
isolating the soldier generation from other aspects and from the vanilla game
setup code, as this triggers the complex namelist mod code.

The main patchfile now introduces an infinite loop after invoking the mutator,
constantly sleeping. The mutator itself is divided into multiple functions,
where each function sets a timer to invoke the next in the chain. This allows
ticks to occur between phases, and resets the internal loop counter.